tohoku/okada: consolidate inversion scripts
